requirements.txt
streamlit>=1.32.0
langchain>=1.0.0
langchain-classic>=1.0.0
langchain-community>=0.4.0
langchain-openai>=1.0.0
langchain-text-splitters>=1.0.0
langchain-core>=1.0.0
chromadb>=1.0.0
openai>=1.14.0
pypdf>=4.1.0
python-dotenv>=1.0.1
tiktoken>=0.6.0
